The basics of async and Readme

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -6,7 +6,7 @@ import asyncio
 from random import uniform
 
 
-async def wait_random(max_delay: float = 10.0) -> float:
+async def wait_random(max_delay: int = 10) -> float:
     '''
     Write an asynchronous coroutine that takes in an integer argument/
     (max_delay, with a default value of 10) named wait_random that /
